Move recommendations module into separate directory
