+++ src/app.js
@@ -1,13 +1,14 @@
-const express       = require("express");
-const app           = express();
-const bodyParser    = require("body-parser");
-const cors          = require('cors')
+const express           = require("express");
+const app               = express();
+const bodyParser        = require("body-parser");
+const cors              = require('cors')
 
-const models        = require("./sequelize");
-const user          = require("./components/user")(models);
-const login         = require("./components/login")(models);
-const activity      = require("./components/activity")(models);
-const quality       = require("./components/quality")(models);
+const models            = require("./sequelize");
+const user              = require("./components/user")(models);
+const login             = require("./components/login")(models);
+const activity          = require("./components/activity")(models);
+const quality           = require("./components/quality")(models);
+const recommendation    = require("./components/recommendation")(models);
 
 //  Add middleware for cors, json parsing, etc.
 app.use(cors());
@@ -17,6 +18,7 @@ app.use("/api/users", user);
 app.use("/api/login", login);
 app.use("/api/activities", activity);
 app.use("/api/qualities", quality);
+app.use("/api/recommendations", recommendation);
 app.use("/api/test", (req, res, next) => {
     res.status(200).send('Hello world!');
 });
