+++ src/components/recommendation/recommendationRoutes.js
@@ -1,0 +1,5 @@
+module.exports = (router, recommendations) => {
+    router.get("/:id/employees", (rq, rs, n) => {recommendations.getContactRecommendations(rq, rs, n);});
+
+    return router;
+}
