+++ src/components/user/userRoutes.js
@@ -4,7 +4,6 @@ module.exports = (router, users) => {
     router.put("/:id", (rq, rs, n) => {users.updateProfile(rq, rs, n);});
     router.post("/:id/contact", (rq, rs, n) => {users.contact(rq, rs, n)});
     router.post("", (rq, rs, n) => {users.create(rq, rs, n);});
-    router.get("/:id/recommendations/contacts", (rq, rs, n) => {users.getContactRecommendations(rq, rs, n);});
 
     router.post("/testdata", (rq, rs, n) => {users.insertTestData(rq, rs, n);});
